Reduce thread contention, more immutable classes and extract char handlers
